requirements.txt
python-dateutil==2.8.2
python-dotenv==1.0.0
requests-ntlm==1.2.0
jira==3.5.2
html5lib==1.1
beautifulsoup4==4.12.2
click==8.1.7